+++ apps/admin/next.config.ts
@@ -1,0 +1,34 @@
+/** @format */
+
+import type { NextConfig } from 'next';
+
+const nextConfig: NextConfig = {
+	async headers() {
+		return [
+			{
+				source: '/_next/:path*',
+				headers: [
+					{
+						key: 'Cache-Control',
+						value: 'no-store, no-cache, must-revalidate, proxy-revalidate',
+					},
+				],
+			},
+		];
+	},
+	reactCompiler: true,
+	images: {
+		unoptimized: true,
+	},
+	allowedDevOrigins: ['*.dev.xernerx.com'],
+	turbopack: {
+		rules: {
+			'*.svg': {
+				loaders: ['@svgr/webpack'],
+				as: '*.js',
+			},
+		},
+	},
+};
+
+export default nextConfig;