+++ apps/cdn/src/app/page.tsx
@@ -1,10 +1,6 @@
 /** @format */
 'use client';
 
-import { useSession } from 'next-auth/react';
-
 export default function Home() {
-	const { data: session } = useSession();
-
 	return <div className="">test</div>;
 }