+++ packages/lib/src/models/xernerx/tokens/Api.ts
@@ -7,7 +7,10 @@ const schema = new Schema(
 		id: { type: String, unique: true, required: true },
 		name: { type: String, required: true },
 		owners: { type: [String] },
-		status: { type: String, required: true, default: 'active', enum: ['active', 'suspended', 'pending'] },
+		status: { type: String, required: true, default: 'active', enum: ['active', 'inactive', 'suspended', 'pending'] },
+		permissions: {
+			isXernerx: Boolean, // can fetch data from /secure
+		},
 		botId: { type: String },
 	},
 	{ timestamps: true }