+++ apps/app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # app
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.3.0
 
 ### Minor Changes