+++ apps/auth/CHANGELOG.md
@@ -1,5 +1,17 @@
 # auth
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+    - @xernerx/providers@0.3.0
+    - @xernerx/components@0.2.7
+    - @xernerx/feedback@0.0.8
+    - @xernerx/lib@0.3.1
+    - @xernerx/ui@0.2.3
+
 ## 0.5.1
 
 ### Patch Changes