+++ apps/bot/CHANGELOG.md
@@ -1,5 +1,11 @@
 # bot
 
+## 1.1.30
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.1.29
 
 ### Patch Changes