+++ apps/cdn/CHANGELOG.md
@@ -1,5 +1,16 @@
 # cdn
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+    - @xernerx/providers@0.3.0
+    - @xernerx/components@0.2.7
+    - @xernerx/feedback@0.0.8
+    - @xernerx/lib@0.3.1
+
 ## 0.4.4
 
 ### Patch Changes