+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @xernerx/lib
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.3.0
 
 ### Minor Changes