+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @xernerx/ui
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.2
 
 ### Patch Changes